code correction of docrows
added the key attribute to each row so that they are updated individually instead of all of them at once when there isn't a key to identify the changed item.
Also removed the state value 'refresher' that used to refresh the whole page instead of only one document row (not even all the documentsrows but the whole page!). boosting significantly the performance of the website